requirements.txt
requests
beautifulsoup4 
langchain 
chromadb 
langchain-chroma
langchain-ollama
PyPDF2
tk
ollama
sentence-transformers 
langchain_huggingface 
nltk